Splitting ear go script into two - image push + compute version
