+++ aqfer/scripts/push-docker-image/push-docker-image.go
@@ -17,41 +17,16 @@ import (
 	"strings"
 	"io"
 	"github.com/aws/aws-sdk-go/aws/awserr"
-	"sort"
-	"strconv"
-	"regexp"
 )
 
 func usage() {
-	fmt.Fprint(os.Stderr, "Usage aq-push-docker-image docker-image-name aws-region [ecr-repository]")
+	fmt.Fprintf(os.Stderr, "Usage aq-push-docker-image docker-image-name version-to-push aws-region [ecr-repository]")
 	os.Exit(1)
 }
 
 func fail(err error) {
-	fmt.Fprint(os.Stderr, "error: %s", err.Error())
+	fmt.Fprintf(os.Stderr, "error: %s", err.Error())
 	os.Exit(2)
-}
-
-func versionSort(versions sort.StringSlice) sort.StringSlice {
-	sort.SliceStable(versions, func(i, j int) bool {
-		a := strings.Split(versions[i], ".")
-		b := strings.Split(versions[j], ".")
-		for k, ak := range a {
-			if k >= len(b) {
-				return true
-			}
-			x, _ := strconv.Atoi(ak)
-			y, _ := strconv.Atoi(b[k])
-			if x > y {
-				return true
-			}
-			if x < y {
-				return false
-			}
-		}
-		return false
-	})
-	return versions
 }
 
 func main() {
@@ -60,7 +35,7 @@ func main() {
 	}
 
 	imageName := os.Args[1]
-	majorVersion := os.Args[2]
+	nextVersion := os.Args[2]
 	awsRegion := os.Args[3]
 
 	var repoName string
@@ -77,34 +52,6 @@ func main() {
 	}
 
 	ecrClient := ecr.New(sess)
-
-	lii := &ecr.ListImagesInput{RepositoryName: aws.String(repoName), Filter:&ecr.ListImagesFilter{TagStatus: aws.String("TAGGED")}}
-	lio, err := ecrClient.ListImages(lii)
-	if err != nil {
-		fail(fmt.Errorf("error listing images on repository: %s", err.Error()))
-	}
-
-	var tags []string
-	versionPattern := regexp.MustCompile(fmt.Sprintf(`\A%s\.\d+\.\d+\z`, majorVersion))
-
-	for _, img :=range lio.ImageIds {
-		tag := *img.ImageTag
-		if versionPattern.MatchString(tag) {
-			tags = append(tags, tag)
-		}
-	}
-	tags = versionSort(tags)
-	var nextVersion string
-	if len(tags) == 0 {
-		nextVersion = "0.0.0"
-	} else {
-		last := tags[0]
-		s := strings.Split(last, ".")
-		v, _ := strconv.Atoi(s[1])
-		s[1] = strconv.Itoa(v + 1)
-		nextVersion = strings.Join(s, ".")
-	}
-
 	log.Printf("building version %s", nextVersion)
 
 	var repo *ecr.Repository
